Set mclgpu_ as a pointer to MCLGPU object
Since it is a pointer, no constructor is needed before the body of the
MCL constructor is called. Thus in the MCL constructor body, we can
initialize mclgpu_ using

 mclgpu_ = new MCLGPU(num_particles)

instead of

  mclgpu = MCLGPU(num_particles)

in which case the destructor of MCLGPU will be called and the allocated
device memory for d_particles_ and d_weights_ will be freed. Thus there
will be an error if we call cudaMemcpy on these pointers.