Init particles around (0,0,0)
initialize_global() will spread particles all over the map. In most
cases, none of the particles is close to the inital pose. Thus the
particles will not follow the car.
Therefore, we initialize the particles around (0,0,0), which is the
default initial pose of the car when the simulator starts. In this way,
it is guaranteed that the particles will following the car.