MCL_cpu() Sync between main and worker threads using condition variable
t_cpu_update() serves as the main thread. It sets up flags, partition the work for
worker threads, and notify_all on the condition variable.
cpu_update() serves as the worker thread function. It does not terminate unless
ros is shutdown.
For details about the sync mechanism, check the comments.

Other changes:
- join threads in the dtor of MCL
- Added initiaize_mt() to initialize threads and set their CPU affinity

TODO
- set scheduling policy and priority is not permitted
- Work on the GPU thread in MCL_hybrid()