Finished multi-threading implementation of MCL_hybrid()
Bound gpu_update() as the gpu thread function. In MCL_hybrid(), the main thread
syncs with both cpu workers and the gpu worker using two different condition
variables.
In MCL_gpu(), mclgpu->update() is called directly.

I also made a lot of effort to make the termination of the multi-threaded program
work. What we need to do is to make sure the main thread as well as all the worker
threads can terminate whenever a SIGINT signal is received.

Check the comments for more implementation details.